pyproject.toml
[project]
name = "modelspec"
version = "0.1.0"
description = "An open-source model intelligence platform. Catalog every AI model, map it to hardware, rank what's best for you."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "falkordb>=1.0",
    "httpx>=0.27",
    "rich>=13.0",
    "typer>=0.12",
    "fastapi>=0.115",
    "uvicorn>=0.30",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
    "ruff>=0.6",
    "mypy>=1.11",
]
researcher = [
    "beautifulsoup4>=4.12",
    "celery>=5.4",
    "anthropic>=0.39",
]

[project.scripts]
modelspec = "cli.modelspec.cli:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.mypy]
python_version = "3.11"
strict = true
